+++ fab/internal/shared.js
@@ -13,39 +13,41 @@ import { classMap } from 'lit/directives/class-map.js'
 import { requestUpdateOnAriaChange } from '../../internal/aria/delegate.js'
 // tslint:disable-next-line:enforce-comments-on-exported-symbols
 export class SharedFab extends LitElement {
-    static properties = {
-        size: { type: String, reflect: true },
-        label: { type: String },
-        lowered: { type: Boolean },
-        href: { type: String },
-    }
-    constructor() {
-        super(...arguments)
-        /**
-         * The size of the FAB.
-         *
-         * NOTE: Branded FABs cannot be sized to `small`, and Extended FABs do not
-         * have different sizes.
-         */
-        this.size = 'medium'
-        /**
-         * The text to display on the FAB.
-         */
-        this.label = ''
-        /**
-         * Lowers the FAB's elevation.
-         */
-        this.lowered = false
-        /**
-         * Sets the underlying `HTMLAnchorElement`'s `href` resource attribute.
-         */
-        this.href = ''
-    }
-    render() {
-        // Needed for closure conformance
-        const { ariaLabel } = this
-        const tag = this.href ? literal`div` : literal`button`
-        return staticHtml`
+  static properties = {
+    size: { type: String, reflect: true },
+    label: { type: String },
+    lowered: { type: Boolean },
+    href: { type: String },
+    expanded: false,
+  }
+  constructor() {
+    super(...arguments)
+    /**
+     * The size of the FAB.
+     */
+    this.size = 'medium'
+    /**
+     * The text to display on the FAB.
+     */
+    this.label = ''
+    /**
+     * Lowers the FAB's elevation.
+     */
+    this.lowered = false
+    /**
+     * Sets the underlying `HTMLAnchorElement`'s `href` resource attribute.
+     */
+    this.href = ''
+    /**
+     * This will allow the fab to expand and use the label
+     */
+    this.expanded = false
+  }
+  render() {
+    // Needed for closure conformance
+    const { ariaLabel } = this
+    const tag = this.href ? literal`div` : literal`button`
+    return staticHtml`
       <${tag}
         class="fab ${classMap(this.getRenderClasses())}"
         aria-label=${ariaLabel || nothing}>
@@ -58,52 +60,49 @@ export class SharedFab extends LitElement {
         ${this.href && this.renderLink()}
       </${tag}>
     `
+  }
+  getRenderClasses() {
+    const isExtended = this.label && this.expanded
+    return {
+      lowered: this.lowered,
+      small: this.size === 'small' && !isExtended,
+      large: this.size === 'large' && !isExtended,
+      extended: isExtended,
     }
-    getRenderClasses() {
-        const isExtended = !!this.label
-        return {
-            'lowered': this.lowered,
-            'small': this.size === 'small' && !isExtended,
-            'large': this.size === 'large' && !isExtended,
-            'extended': isExtended,
-        }
-    }
-    renderTouchTarget() {
-        return html`<div class="touch-target"></div>`
-    }
-    renderLabel() {
-        return this.label ? html`<span class="label">${this.label}</span>` : ''
-    }
-    renderIcon() {
-        const { ariaLabel } = this
-        return html`<span class="icon">
-      <slot
-        name="icon"
-        aria-hidden=${ariaLabel || this.label
-                ? 'true'
-                : nothing}>
+  }
+  renderTouchTarget() {
+    return html`<div class="touch-target"></div>`
+  }
+  renderLabel() {
+    return this.expanded && this.label ? html`<span class="label">${this.label}</span>` : ''
+  }
+  renderIcon() {
+    const { ariaLabel } = this
+    return html`<span class="icon">
+      <slot name="icon" aria-hidden=${ariaLabel || this.label ? 'true' : nothing}>
         <span></span>
       </slot>
     </span>`
-    }
-    renderLink() {
-        // Needed for closure conformance
-        const { ariaLabel } = this
-        return html`
+  }
+  renderLink() {
+    // Needed for closure conformance
+    const { ariaLabel } = this
+    return html`
       <a
         class="link"
         id="link"
         href="${this.href}"
         target="${this.target || nothing}"
-        aria-label="${ariaLabel || nothing}"></a>
+        aria-label="${ariaLabel || nothing}"
+      ></a>
     `
-    }
+  }
 }
-(() => {
-    requestUpdateOnAriaChange(SharedFab)
+;(() => {
+  requestUpdateOnAriaChange(SharedFab)
 })()
 /** @nocollapse */
 SharedFab.shadowRootOptions = {
-    mode: 'open',
-    delegatesFocus: true,
+  mode: 'open',
+  delegatesFocus: true,
 }
